fix :회원정보 수정시 기존 이미지 null 여부 조건 추가

+++ karma/src/main/java/com/ecyce/karma/domain/user/service/UserService.java
@@ -120,7 +120,9 @@ public class UserService {
     /* 사용자 이미지 수정 */
     public void modifyProfileImage(User user, MultipartFile multipartFile) {
       String deletedUrl = userRepository.findProfileImageUrl(user.getUserId());
-      s3Uploader.deleteFile(deletedUrl);
+        if (deletedUrl != null && !deletedUrl.isEmpty()) {
+            s3Uploader.deleteFile(deletedUrl);
+        }
 
       User targetUser = userRepository.findByUserId(user.getUserId());
       String updateUrl = s3Uploader.saveFile(multipartFile);
